+++ src/functions/util/pkgbuild.sh
@@ -283,3 +283,20 @@ get_integlist() {
 		printf "%s\n" "${INTEGRITY_CHECK[@]}"
 	fi
 }
+
+# Get all occurances of a variable plus it's extensions (i.e. for 'depends',
+# 'depends' and 'focal_depends').
+get_extended_variables() {
+	printf '%s\n' "${env_keys[@]}" | grep -F "_${1}" | head -c -1
+	printf '%s\n' "${env_keys[@]}" | grep -F "${1}_" | head -c -1
+
+	if in_array "${1}" "${env_keys[@]}"; then
+		echo "${1}"
+	fi
+}
+
+# Split a dependency by it's pipe character, i.e.
+# 'hi|me' -> 'hi' and 'me' returned.
+split_dep_by_pipe() {
+	echo "${1}" | sed 's/|/\n/g' | head -c -1
+}