+++ CHANGELOG.md
@@ -8,7 +8,9 @@ Note that the `[Unreleased]` section contains all changes that haven't yet made 
 
 ## [Unreleased]
 ### Added
-- Added support for custom installing directories during builds (#61).
+- Add support for custom installing directories during builds (#61).
+- Add support for `|` characters in dependency fields.
+- Add CLI options to print makedeb-styled message (#156).
 
 ## [14.1.2] - 2022-05-07
 ### Fixed